add the artifactManager from the hook and test it
