--- v-next/hardhat-ethers/test/hook-is-set.ts
@@ -1,13 +1,0 @@
-// describe("HRE extensions", function () {
-//   it("should extend hardhat runtime environment", function () {
-//     assert.isDefined(this.env.ethers);
-//     assert.containsAllKeys(this.env.ethers, [
-//       "provider",
-//       "getSigners",
-//       "getImpersonatedSigner",
-//       "getContractFactory",
-//       "getContractAt",
-//       ...Object.keys(ethers),
-//     ]);
-//   });
-// });
